chore: src/ 내부 폴더 구조 초기 설정 #3
