+++ App.js
@@ -1,19 +1,7 @@
 import React from "react";
-import "./global.css";
-import { View, Text, TouchableOpacity } from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
+import "./global.css"
+import MainPage from "./src/screens/MainPage";
 
-export default function MainPage() {
-  return (
-    <SafeAreaView>
-      <View className="flex justify-between w-full max-w-md mt-4">
-        <TouchableOpacity className="bg-orange-500 text-white font-bold py-4 px-12 rounded-2xl shadow">
-          <Text className="text-white font-bold">저장소!</Text>
-        </TouchableOpacity>
-        <TouchableOpacity className="bg-orange-500 text-white font-bold py-4 px-12 rounded-2xl shadow">
-          <Text className="text-white font-bold">설정!</Text>
-        </TouchableOpacity>
-      </View>
-    </SafeAreaView>
-  );
+export default function App() {
+  return <MainPage />;
 }
